research-assistant/requirements.txt
numpy
requests
python-dotenv
feedparser
PyMuPDF
tiktoken
sendgrid
openai
pinecone-client
pytest
